requirements.txt
python-dotenv
assemblyai
langchain
faiss-cpu
langchain_google_genai
langchain_community
streamlit